main: Separate the disassembler into its own file
Makes it easier to locate.
